% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackImages.R
\name{stackImages}
\alias{stackImages}
\alias{stackimages}
\title{stackImages}
\usage{
stackImages(input_dir = NULL, input_files = NULL,
  stackMethod = "average")
}
\arguments{
\item{input_dir}{A string}

\item{input_files}{A list of strings}

\item{stackMethod}{A string}
}
\description{
Function to stack given images
}
\details{
The input files should be tif-format.
The output will be a tif-file as well.
}
\author{
Kai Budde
}
